fix: correct alpha validation

+++ supervised_learning/classification/14-neural_network.py
@@ -143,7 +143,7 @@ class NeuralNetwork:
             raise ValueError("iterations must be a positive integer")
 
         # Validate alpha
-        if type(alpha) is not int:
+        if type(alpha) is not float:
             raise TypeError("alpha must be a float")
         if alpha <= 0:
             raise ValueError("alpha must be positive")
